Strengthen Ties to USA canada changes
Attempting to rebalance the strengthen ties to usa path so it's not as overpowered as it was but still is usuable. 

Copying the model of containment on UK of destroying your stability at the cost of war support and better economic laws. 
Added more chromium and tungsten as ressources to make it a bit more even.
Change the spirits it gives to hurt civilian economy build up and give them a small buff for fighting in africa.