Navs nerf
Lowered their defense and naval attack a bit so they are less opressive. ATM a full nav carrier can one shot a full fighter carrier and thats just not right. Also navs overdecides the winner of a battle and hopefully this helps tone it down. More changes will be done if needed.