5 starting research slots for italy at start
They have to tech navy, infantry, air and sometimes even tanks, but only starts with 4 tech slots which is no enough. You can get 6 but it's too late into the game imo and makes it a little too hard for italy to get techs it needs compared to uk that can get 7 techs slots 2 focuses in.