Supremacy rebalance
Missions that do not require taking any risks should not give nearly as much supremacy as the missions that do. A navy afk in its fort on naval invasion support or strike force is not taking any risk what so ever and should not be rewarded with bonus supremacy. Using patrol and control escort missions should be rewarded as the fleet has to take a risk and be active in the zone, making it possible to attack it, so it should be the missions that are encouraged and rewarded in order to get more supremacy in a sea zone